requirements.txt
requests>=2.28.0
websockets>=12.0
pymysql>=1.1.0
redis>=5.0.0
celery>=5.3.0
PyYAML>=6.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
